+++ src/js/core.js
@@ -561,7 +561,7 @@ $(document).ready(function ($) {
 
                             // Update Status and Quality
                             rSearchTerm = /(\w+)\s\((.+?)\)/;
-                            htmlContent = ep.status.replace(rSearchTerm, "$1" + ' <span class="quality ' + ep.quality + '">' + "$2" + '</span>');
+                            htmlContent = ep.status.replace(rSearchTerm, "$1" + ' <span class="badge text-white ' + ep.quality + '">' + "$2" + '</span>');
                             parent.closest('tr').prop("class", ep.overview + " season-" + ep.season + " seasonstyle font-weight-bold text-dark");
                         }
                         // update the status column if it exists
@@ -632,7 +632,7 @@ $(document).ready(function ($) {
                         }
                         // applying the quality class
                         var rSearchTerm = /(\w+)\s\((.+?)\)/;
-                        htmlContent = data.result.replace(rSearchTerm, "$1" + ' <span class="quality ' + data.quality + '">' + "$2" + '</span>');
+                        htmlContent = data.result.replace(rSearchTerm, "$1" + ' <span class="badge text-white ' + data.quality + '">' + "$2" + '</span>');
                         // update the status column if it exists
                         parent.siblings('.col-status').html(htmlContent);
                         // Only if the queuing was successful, disable the onClick event of the loading image